Add textures to objects
The MaterialManager holds all the texture configurations and applies
textures to objects by material. Textures will be cached so that the same
texture is not recreated over and over again. The material manager is also
capable of restoring the original state again.

The texture configuration can be set via the task panel when double
clicking the TextureConfig object.