Add support for rotation of individual faces
The mapping method could not figure out the right rotation for each
individual face because it depends heavily on the texture itself and we
can't inspect the texture programmatically. So we give the user the power
to rotate the texture for individual faces.

The face configurations will not be stored by name because then we would
be affected by the topological naming issue and we don't have face names
when mapping textures, because we work on the coin scene graph directly.
So we store the vertices of a face and compare them later on. This should
be pretty relyable.

To match faces by vertices we have to account for the SoTransform node and
normalize all vertices by the transformation.

This commit adds support for rotating textures on faces:
 - Adds possibility to load and save face overrides containing a rotation
   angle for each texture config
 - Faces can check if they match a face override by comparing the original
   vectors with the vectors stored inside a face override
 - Utitlity to select faces and the vectors
 - Adds command to configure rotation for individual faces

fixes #10