Fix missing FlatLines
Previously the texture was set on the root node of the object. This means
the texture was also applied to the wireframe part. This is not what we
want.

This commit fixes this and applies the texture only to the shaded node of
the object. So it only affects the faces of the object.

fixes #15