Fix calculation of 360 panorama texture
Previously the calculation was way to simple. It did not take into account,
that the start of the image should always be at the positive y axis
so that north always faces up.

This commit fixes the calculation for all environments where the planes
overshoot the x and y axis.