Remove texmodel.REPLACE
This property disabled phong shading and resulted in a pretty flat
look when textures where applied.

Now we set the diffuse color of the object to white so that phong shading
is applied, but the color is only taken from the texture. When the
textures are removed we reset the color to the original one.

fixes #26