Add Bump Map support to TextureManager
When a bump map is configured in the TextureConfig the Texture manager
adds the following nodes to the scenegraph to get bump mapping applied
 - SoBumpMap with the bump map image (normal or height map)
 - The texture is encapsulated in a SoTextureUnit with the value of 1.
   Otherwise the bump map would be also used as color texture.
 - The SoBumpMap also gets the SoTextureCoordiante2 node applied.

On cleanup all available nodes added by the texture manager will be
removed again. Now we check if a node was really added, because bumpmaps
are completely optional.

references #8