Fix potential None shaded node
When loading a document the shaded node could be None, when the
TextureConfig gets executed before the whole document is set up.

To prevent throwing an exception we check for None and print a message
in such a case. That would make it easier to trace down such
problems later on, when there might be some real objects without shaded
node.