Add textures to the renderer workbench
This commit extends the TextureConfig to support the new Material
Pipeline of the Renderer Workbench. When a texture config is added
to the material pipeline, it will enhance the material and set the
texturing releveant settings.

To make this possible some existing stuff was refactored, to also support
calculating texture coordinates for FreeCAD Meshes.